[feature] kotlin

+++ applib/src/main/java/com/king/applib/util/Exts.kt
@@ -12,8 +12,8 @@ import android.util.TypedValue
 
 val Float.dp
     get() = TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, this,
-        Resources.getSystem().displayMetrics)
+        Resources.getSystem().displayMetrics).toInt()
 
 val Float.sp
     get() = TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP, this,
-        Resources.getSystem().displayMetrics)
+        Resources.getSystem().displayMetrics).toInt()
